requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
azure-core==1.35.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.23.1
azure-monitor-opentelemetry==1.6.10
azure-monitor-opentelemetry-exporter==1.0.0b40
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
Deprecated==1.2.18
fastapi==0.116.1
firebase-admin==6.9.0
fixedint==0.1.6
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.1.1
msrest==0.7.1
oauthlib==3.3.1
opentelemetry-api==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-dbapi==0.49b2
opentelemetry-instrumentation-django==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-flask==0.49b2
opentelemetry-instrumentation-psycopg2==0.49b2
opentelemetry-instrumentation-requests==0.49b2
opentelemetry-instrumentation-urllib==0.49b2
opentelemetry-instrumentation-urllib3==0.49b2
opentelemetry-instrumentation-wsgi==0.49b2
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.11.0
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyodbc==5.2.0
pyparsing==3.2.3
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wrapt==1.17.2
zipp==3.23.0
